requirements.txt
art==5.8
boto3==1.26.35
botocore==1.29.35
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==2.1.1
cryptography==38.0.4
ecdsa==0.18.0
envs==1.4
idna==3.4
jmespath==1.0.1
pyasn1==0.4.8
pycognito==2022.11.2
pycparser==2.21
PyJWT==2.6.0
python-dateutil==2.8.2
python-jose==3.3.0
requests==2.28.1
rsa==4.9
s3transfer==0.6.0
six==1.16.0
termcolor==2.2.0
urllib3==1.26.13
